// MeskChatClient/src/types/responseEntityOfLoginCommandResponse.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * MeskChatApplication.WebApi | v1
 * OpenAPI spec version: 1.0.0
 */
import type { LoginCommandResponse } from './loginCommandResponse';
import type { ResponseEntityOfLoginCommandResponseValidationErrors } from './responseEntityOfLoginCommandResponseValidationErrors';

export interface ResponseEntityOfLoginCommandResponse {
  isSuccess?: boolean;
  /** @nullable */
  message?: string | null;
  data?: LoginCommandResponse;
  statusCode?: number;
  /** @nullable */
  validationErrors?: ResponseEntityOfLoginCommandResponseValidationErrors;
}